Linear Regression Implemented

+++ app/src/main/java/johan/run_hub/db/entities/Exercise.kt
@@ -4,12 +4,12 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity(tableName = "exercise_table")
-data class Exercise (
+data class Exercise(
     var distanceCovered: Int = 0,
     var elapsedTimeInMillis: Long = 0L,
     var exerciseDate: Long = 0L,
     var averageSpeed: Float = 0f,
-    var caloriesBurned: Int = 0,
+    var caloriesBurned: Double = 0.0,
     var exerciseType: String = ""
 ) {
     @PrimaryKey(autoGenerate = true)
